Unit test and random generator improvement for accurate generation of correct probabilities.

+++ SimplifiedSlotMachine/SimplifiedSlotMachine.DataModel/Symbol.cs
@@ -5,10 +5,10 @@
         public string Name { get; set; }
         public string Letter { get; set; }
         public decimal Coefficient { get; set; }
-        public decimal Probability { get; set; }
+        public double Probability { get; set; }
         public bool IsWildcard {  get; set; }
 
-        public Symbol(string name, string letter, decimal coefficient, decimal probability, bool isWildcard = false)
+        public Symbol(string name, string letter, decimal coefficient, double probability, bool isWildcard = false)
         {
             this.Name = name;
             this.Letter = letter;
